move: Move InternalUtils to "internal" package.
