+++ src/main/java/jp/co/moneyforward/autotest/framework/action/Scene.java
@@ -3,7 +3,7 @@ package jp.co.moneyforward.autotest.framework.action;
 import com.github.dakusui.actionunit.core.Action;
 import com.github.valid8j.pcond.fluent.Statement;
 import jp.co.moneyforward.autotest.actions.web.Value;
-import jp.co.moneyforward.autotest.framework.utils.InternalUtils;
+import jp.co.moneyforward.autotest.framework.internal.InternalUtils;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -16,7 +16,7 @@ import static com.github.dakusui.actionunit.core.ActionSupport.sequential;
 import static com.github.valid8j.classic.Requires.requireNonNull;
 import static java.util.Arrays.asList;
 import static jp.co.moneyforward.autotest.framework.action.AutotestSupport.*;
-import static jp.co.moneyforward.autotest.framework.utils.InternalUtils.simpleClassNameOf;
+import static jp.co.moneyforward.autotest.framework.internal.InternalUtils.simpleClassNameOf;
 
 ///
 /// An interface that represents a reusable unit of an action in **insdog**'s programming model.
