+++ src/main/java/jp/co/moneyforward/autotest/framework/cli/CliUtils.java
@@ -3,7 +3,7 @@ package jp.co.moneyforward.autotest.framework.cli;
 import com.github.valid8j.pcond.forms.Predicates;
 import jp.co.moneyforward.autotest.framework.annotations.AutotestExecution;
 import jp.co.moneyforward.autotest.framework.testengine.AutotestEngine;
-import jp.co.moneyforward.autotest.framework.utils.InternalUtils;
+import jp.co.moneyforward.autotest.framework.internal.InternalUtils;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Tags;
 import org.junit.platform.engine.TestDescriptor;
@@ -33,8 +33,8 @@ import static com.github.valid8j.classic.Requires.requireNonNull;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toMap;
 import static jp.co.moneyforward.autotest.framework.core.ExecutionEnvironment.testResultDirectoryFor;
-import static jp.co.moneyforward.autotest.framework.utils.InternalUtils.removeFile;
-import static jp.co.moneyforward.autotest.framework.utils.InternalUtils.writeTo;
+import static jp.co.moneyforward.autotest.framework.internal.InternalUtils.removeFile;
+import static jp.co.moneyforward.autotest.framework.internal.InternalUtils.writeTo;
 import static org.junit.platform.engine.discovery.DiscoverySelectors.selectClass;
 import static org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder.request;
